requirements.txt
PyMuPDF==1.22.5
beautifulsoup4==4.12.2
openai==0.27.8
psycopg2-binary==2.9.8
transformers==4.33.2
torch==2.1.0
accelerate==0.21.0
